1559: Move new build section utility to common
