+++ accessibility_monitoring_platform/apps/common/templates/common/view_toc_link.html
@@ -1,10 +1,10 @@
 {% if section.anchor %}
     <a href="#{{ section.anchor }}" class="govuk-link govuk-link--no-visited-state">
-        {{ section.name }}{% if section.complete %} <span class="govuk-visually-hidden">complete</span>{% endif %}</a>
+        {{ section.name }}{% if section.complete %}<span class="govuk-visually-hidden">complete</span>{% endif %}</a>
 {% else %}
-    {{ section.name }}{% if section.complete %} <span class="govuk-visually-hidden">complete</span>{% endif %}
+    {{ section.name }}{% if section.complete %}<span class="govuk-visually-hidden">complete</span>{% endif %}
 {% endif %}
 |
 <a id="{{ section.edit_url_id }}" href="{{ section.edit_url }}" class="govuk-link govuk-link--no-visited-state">
-    Edit{% if section.complete %} <span class="govuk-visually-hidden">complete</span>{% endif %}</a>
+    Edit{% if section.complete %}<span class="govuk-visually-hidden">complete</span>{% endif %}</a>
 {% if section.complete %}&check;{% endif %}
